+++ ecosystem/typescript/sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to the Aptos Node SDK will be captured in this file. This ch
 
 ## Unreleased
 
+N/A
+
+## 1.3.11 (2022-08-31)
+
 - Upgraded typescript version from 4.7.4 to 4.8.2, as well as linter package versions.
 - **[Breaking Change]** ModuleBundle transaction support is removed. Instead, SDK users should use `AptosClient.publishPackage` to publish Move packages.
 - Expose detailed API errors.